OP-2669 - better selection of work folder
Because work folder in Harmony is local, it must be replaced with folder where workfile will be finally published. In some cases naming convention of expected files didn't match correct value, which resulted in not updating the path.
This explicitly replaces dir name of expected file with staging folder where renders will be created on a DL machine.