Expanding the scope of mvLook to publish mipmapped textures too, like TDL's generated by 3delight. During collection, we check to see if the TDL's exist beside the original texture, and if they do they get added to the collection.
Validation of these will depend on the publish intent and whether
`expectMipMap` is True in the publish set.

If mipmaps are expected and the intent is correct, the mipmaps will be
published along with the original file, which will accelerate rendering,
since 3dl will find that TDL and use it instead of having to generate it.